+++ js/util.js
@@ -1,0 +1,11 @@
+const getRandomInteger = (min, max) => {
+  const result = Math.random() * (max - min + 1) + min;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => {
+  const element = elements[getRandomInteger(0, elements.length - 1)];
+  return element;
+};
+
+export {getRandomInteger, getRandomArrayElement};
